Show individual page reports
Show accessibility reports for app pages.

You can find the list of reports at /front-end/accessibility
