+++ app/controllers/fenrir_view/accessibility_controller.rb
@@ -5,5 +5,11 @@ module FenrirView
     def index
       @content = FenrirView::AccessibilityFacade.new
     end
+
+    def page
+      @content = FenrirView::PageAccessibilityFacade.new(page: params[:page])
+
+      return render_404 unless @content.mobile_report? || @content.desktop_report?
+    end
   end
 end
