+++ lib/fenrir_view.rb
@@ -5,6 +5,7 @@ require 'fenrir_view/component/accessibility_issue'
 require 'fenrir_view/component/accessibility_overview'
 require 'fenrir_view/component/accessibility_screenshot'
 require 'fenrir_view/component/accessibility'
+require 'fenrir_view/page/accessibility'
 require 'fenrir_view/component/examples'
 require 'fenrir_view/component/health'
 require 'fenrir_view/configuration'
